requirements.txt
selenium==3.141.0
Flask==2.0.0
Pillow==8.3.1
